+++ CHANGE_LOG.md
@@ -1,8 +1,14 @@
+- [v2.1.2](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v2.1.2)
 - [v2.1.1](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v2.1.1)
 - [v2.1.0](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v2.1.0)
 - [v2.0.0](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v2.0.0)
 - [v1.0.1](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v1.0.1)
 - [v1.0.0](https://github.com/joaomcteixeira/Tauren-MD/blob/master/CHANGE_LOG.md#v1.0.0)
+
+# v2.1.2
+
+- Added data export function `communicate.export.save_data_array_to_file()`
+  - now each plotting routine exports used data in .csv files.
 
 # v2.1.1